refactor: move JS and CSS into App.vue
As a first step I start with the root component by folding the .js and
.css into the .vue file. The content was then formatted by ESLint.

Future refactorings would rewrite the structure to follow the
Composition API.

By breaking the work into smaller steps I'm hopefully able to explain
what is going on in more nuance.
