--- src/App.css
@@ -1,161 +1,0 @@
-:root {
-  --yellow: #d8b33d;
-  --green: #5f8e5d;
-  --red: #cd4e49;
-  --offwhite: #dfd6b3;
-
-  --dark-grey: #444;
-
-  --unlighterer-grey: #aaa;
-  --unlighter-grey: #bbb;
-  --unlight-grey: #ddd;
-  --light-grey: #eee;
-
-  /* Using this as opposed to 0.25em/0.25rem to avoid rounding
-	   weirdness that creates uneven gaps */
-  --small-margin: 4px;
-
-  --font-stack: wakamai-fondue, notdef;
-  --system-font-stack: system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
-    Roboto, Ubuntu, "Helvetica Neue", Oxygen, Cantarell, sans-serif;
-}
-
-@font-face {
-  font-family: notdef;
-  src: url("assets/AND-Regular.woff2");
-}
-
-html {
-  box-sizing: border-box;
-  min-height: 320px;
-  min-width: 320px;
-  font-family: var(--system-font-stack);
-  scroll-behavior: smooth;
-}
-
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
-}
-
-body,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-ol,
-ul {
-  margin: 0;
-  padding: 0;
-  font-weight: normal;
-}
-
-ol,
-ul {
-  list-style: none;
-}
-
-table {
-  font-family: inherit;
-  border-collapse: collapse;
-  border-spacing: 0;
-}
-
-select,
-button {
-  font-family: inherit;
-  font-size: inherit;
-}
-
-a {
-  color: inherit;
-}
-
-/* Global styles */
-h2 {
-  font-size: 1.5rem;
-  padding: 2rem 0;
-  margin: 4rem 0;
-  background: var(--green);
-  color: white;
-  text-align: center;
-}
-
-h3 {
-  color: var(--red);
-  font-weight: bold;
-  font-size: 1.25rem;
-  margin: 2rem 0 1rem 0;
-}
-
-.content {
-  margin-left: auto;
-  margin-right: auto;
-  max-width: 68rem;
-  padding: 0 1rem;
-}
-
-.opentype-label {
-  font-family: monospace;
-  border-radius: 0.2rem;
-  display: inline-flex;
-  align-items: center;
-  height: 1.5rem;
-  padding: 0 0.75rem;
-  margin-right: var(--small-margin);
-  background: var(--offwhite);
-}
-
-.button {
-  padding: 0.25rem 1rem;
-  border-radius: 99999px;
-  margin: 0;
-  border: 0;
-  background: none;
-  font-size: 0.85rem;
-  background: var(--unlighter-grey);
-  cursor: pointer;
-  min-width: 6em;
-  user-select: none;
-  color: black;
-}
-
-.button:hover {
-  box-shadow: inset 0 0 999px rgba(0, 0, 0, 0.15);
-}
-
-.button[disabled] {
-  background: var(--unlighterer-grey);
-}
-
-.button.active {
-  background: var(--dark-grey);
-  color: white;
-}
-
-.button.active:hover {
-  box-shadow: inset 0 0 999px rgba(0, 0, 0, 0.4);
-}
-
-.button.on {
-  background: var(--green);
-  color: white;
-}
-
-.button.off {
-  background: var(--red);
-  color: white;
-}
-
-@media screen and (prefers-reduced-motion: reduce) {
-  *,
-  *:before,
-  *:after {
-    animation: none !important;
-    transition: none !important;
-  }
-}
